cloudrun/river-ml-api/app/requirements.txt
Flask
gunicorn
google-cloud-pubsub
google-cloud-storage
google.cloud
river